requirements.txt
flask==2.3.3
flask-cors==4.0.0
flask-session==0.5.0
pandas==2.1.1
pyodbc==4.0.39
duckdb==0.9.0
bcrypt==4.0.1
python-dotenv==1.0.0
Werkzeug==2.3.7
openpyxl==3.1.2
azure-identity==1.14.0
azure-storage-file-datalake==12.12.0
redis==5.0.1
